feat: add explain TableFunction for substrait
This adds a new table function, `explain_substrait`, as well as a new
execution kernel for the function and binding kernel. This reuses the
`FromSubstraitFunctionData` structure since it contains all the pieces
we would need.

In a rebase, we re-add `FromSubstraitFunctionData` and we allow
ourselves to rename some functions since others we were matching the
style of have been removed. Changes upstream changed TableFunctions to
return a TableRef instead of a QueryResult, but to return an explain
plan I believe we still need to return a QueryResult (unless we
construct and return an `ExplainRelation` and call `GetTableRef` on
that).