fix(bad-rebase): fixed typos and added test
I had to re-add FromSubstraitFunctionData, so in the process I forgot to
return unique_ptr<FunctionData> instead.

Also re-added missing closing brace.

This also adds a test, but I haven't been able to actually test it.
Additionally, it seems that calling explain is truncating the result, so
I need to figure out how to prevent that from happening in order to
create a proper expected output (and to make the function itself
useful).